Extract multi-dimensional array compiler fixtures
